requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
redis==4.4.0
celery==5.2.7
djangorestframework==3.13.1
django-redis==5.2.0
django-celery-beat==2.2.0
requests==2.26.0
djangorestframework-simplejwt==5.2.2
pyjwt==2.7.0
django-debug-toolbar==4.1.0
django-filter==23.2